pyproject.toml
[project]
name = "pygram"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.2",
    "django-filter>=25.1",
    "django-rest-framework-nested>=0.0.1",
    "djangorestframework>=3.16.0",
    "markdown>=3.8",
    "pillow>=11.2.1",
]
